Move SageAreaModifier to its own file
